Add state-update application
Adds the Host state-update service to support updating Dbus property
CurrentHostStatus base on status of gpio pin S0_FW_BOOT_OK. When SCP boots up
sucessfully, it will pull S0_FW_BOOT_OK to high. This indicates that
the host firmware is running.

The idea is whenever the service is started or Dbus property CurrentHostStatus
is changed or CurrentPowerStatus is changed to On, the service will verify
the value of GPIO pin S0_FW_BOOT_OK and value of CurrentHostStatus. If they
mismatch, the service will update CurrentHostStatus base on the status
of GPIO pin S0_FW_BOOT_OK.

Tested:
1. TC1 - The Host is ON:
 - Power off the host.
 - Check CurrentHostState, it should be Off.
 - Power on the host.
 - Check CurrentHostState, it should be Running.

2. TC2 - The Host is OFF:
 - Power on the host.
 - Check CurrentHostState, it should be Running.
 - Power off the host.
 - Check CurrentHostState, it should be Off.

3. TC3 - Host is Running.
 - Reboot BMC. After BMC boots up.
 - Check CurrentHostState, it should be Running.
 - Repeat TC1

4. TC4 - Host is Off.
 - Reboot BMC. After BMC boot up.
 - Check CurrentHostState, it should be Off.
 - Repeat TC2.

5. TC5 - Host is On.
 - Check CurrentHostState, it should be Running.
 - Call busctr set-property to set CurrentHostState to Off
 - Check CurrentHostState, it should be Running.
   This is the invalid change from user.

6. TC6 - Host is On.
 - Check CurrentHostState, it should be Running.
 - In host linux, call Shutdown.
 - Check CurrentHostState, it should be Off.

7. TC7 - Host is On.
 - Check CurrentHostState, it should be Running.
 - In host linux, call reboot.
 - After the host boot up.
 - Check CurrentHostState, it should be Running.

8. TC8 - Host is On.
 - Check CurrentHostState, it should be Running.
 - Set power policy to Always-on or Always-off
 - AC power. After BMC boot up.
 - The CurrentHostState should match with the host state.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>