driver-binder: Only uses the changed signal of CurrentHostState to bind/unbind
Phosphor-state-manager supports updating the transitioning states to the
CurrentHostState dbus property of xyz.openbmc_project.State.Host interface.
So using the propertyChanged signal of CurrentHostState is enough to
bind/unbind the SMPro drivers. Using CurrentPowerState is redundant.
This commit removes matching propertyChanged signal of CurrentPowerState in
driver binder.
Moreover, driver-binder detects the running state of the host by maching
the "Running" string at the end of the CurrentHostState state.
After support updating the transitioning states to the CurrentHostState,
Phosphor-state-manager updates "..Host.HostState.TransitioningToRunning"
string to CurrentHostState when the host state is transitioning to On.
This causes the driver-bind binds the SMPro-drivers before the host is already
On. This commit corrects the matching string from "Running" to ".Running".

Tested:
	1. Power off/on/reset/cycle/soft the host.
	2. The SMPro-drivers should be unbinded before the host is off/reboot.
	3. They shoud be binded after the host is already running.

Signed-off-by: Thu Ba Nguyen <tbnguyen@amperecomputing.com>