[Workaround] Asserting JTAG_TRST on PCP Power Up for AmpereOne Ax silicon
Due to a subset of the JTAG logic residing in the Processor Complex
(PCP) power domain, it is possible for that logic to come out of reset
in an improper state.  JTAG logic is typically reset via the JTAG TRST
signal, but the TRST signal will not take effect on JTAG logic in the
PCP domain if applied as part of the system reset.  For the TRST to
propagate to all JTAG logic in the SOC, it must be applied after the
PCP power comes up.

This commits implements the process ampere_trst_wa which monitors the
SoC's Sx_PCP_PWRGD signal and pulse the JTAG_TRST for 1ms.

Tested:
1. Verify the ampere_trst_wa utility with the test system.
2. The host should boot normally.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>