mctp-ctrl: bind services to pldmd
When pldmd service is restarted with two EIDs 20 and 22 in mctp-ctrl
D-Bus objects, it will discover two sockets at the same time. Because
PLDM command GetDevPDRs takes much time and workload in SMBus but
there is only one SMBus for two sockets so this can lead the
conflict in accessing the PDRs which causes the time out in
accessing PDRs.
When pldmd is time out in accessing PDRs of one terminus it will
bypass adding the sensors for that terminus because the PDRs of one
terminus have a tight relationship.
So getting PDRs of two sockets at the same time should be avoided to
make sure to prevent the failure in accessing PDRs. This means the
mctp-ctrl should be restarted when pldmd restarts. This commit binds
the mctp-ctrl to pldmd service to support that behavior.

Tested:
1. Restart pldmd service.
2. Mctp-ctrl service should be restarted.

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>