ampere-cpu-present: remove unnescessary codes
CPU dbus' present property identifies the installation status of the
CPU, not the available status. So, this property should not be changed
when the Host transitions to OFF or ON.
This commit removes the redundant codes so that the property correctly
reflects the CPU present status.

Tested:
1. Reboot BMC. Check CPU present D-Bus property.
`busctl tree xyz.openbmc_project.Ampere.Cpu`
2. Power off the host. Check CPU present property.
`busctl tree xyz.openbmc_project.Ampere.Cpu`

Signed-off-by: Thu Nguyen <thu@os.amperecomputing.com>