crash-capture: avoid conflict SPINOR during BMC booting
When the BMC turns ON the Host via "always-on" power restore policy, the
crash capture service might be started at the same time which switches
the Host SPI-NOR to the BMC to check for BERT data. This causes the
conflict of Host SPI-NOR access.
This commit adds enfore the crash capture service to start before the
phosphor-discover-system-state service to avoid the confliction.

Signed-off-by: Dung Cao <dung@os.amperecomputing.com>