Verify SMPro Hwmon when Host is on
Currently, we verify SMPro Hwmon root path during initialization.
However, when BMC boots with Host state as Off (such as when BMC boots
with always-off policy, or after AC cycle), those hwmon files can't not
be found yet and the program exits with error code. This will cause the
`xyz.openbmc_project.altra_host_error_monitor.service` service to be
repeatedly restarted and might hit the systemd unit restart limit rate,
which can send the service to permanent failure.

This commit will only verify SMPro Hwmon root path on the first time
that Host turns on. If SMPro Hwmon root path is not found in that case,
the program will throw exception.

This also let `initHwmonRootPath()` returns boolean, instead of integer
to simplify and clarify the logic.

Tested:

1. AC cycle with always-on or always-off policy
2. Wait until BMC state is ready
3. Check `xyz.openbmc_project.altra_host_error_monitor.service`
$systemctl status xyz.openbmc_project.altra_host_error_monitor.service

Expect that the service's status is Running

Signed-off-by: Chau Ly <chaul@amperecomputing.com>